+++ apps/decodex/src/state.rs
@@ -6,7 +6,7 @@
 };
 use std::{
 	cmp,
-	collections::{BTreeMap, BTreeSet, HashMap, HashSet},
+	collections::{HashMap, HashSet},
 	fs::{self, File, OpenOptions, TryLockError},
 	io::{ErrorKind, Read, Seek, SeekFrom, Write},
 	path::{Path, PathBuf},
@@ -40,6 +40,8 @@ use crate::{
 	prelude::{Result, eyre},
 	tracker::records::{self, LinearExecutionEventRecord},
 };
+
+mod project_run_recovery;
 
 include!("state/store.rs");
